Merged All versions into one, options are now available to select which features you would like to activate. These are located on the Extra Settings screen.
Also cut out all code that was left over from the original mouse2keyboard. I made it so that the menu option for game selection is now just a reset to cemu.exe, meaning you would need to manually change its setting if your Cemu is called something different, or you want to have it activate another window.
Added my own version number to the about, and top of the settings screen, this should assist me in troubleshooting, as I'll know what version the user is using easier.